+++ make.py
@@ -152,12 +152,6 @@ def ut():
     run_cmd('coverage run --source bingads -m py.test -k "not functional" -v --strict')
     run_cmd('coverage report')
 
-
-@task('run all v11 unit tests under current interpreter, and print coverage report')
-def v11_ut():
-    run_cmd('coverage run --source bingads -m py.test v11tests/ -k "not functional" -v --strict')
-    run_cmd('coverage report')
-	
 @task('run all v12 unit tests under current interpreter, and print coverage report')
 def v12_ut():
     run_cmd('coverage run --source bingads -m py.test v12tests/ -k "not functional" -v --strict')
@@ -167,11 +161,6 @@ def v12_ut():
 @task('run all functional tests under current interpreter.')
 def ft():
     run_cmd('py.test -k "functional" -v --strict')
-
-
-@task('run all v11 functional tests under current interpreter.')
-def v11_ft():
-    run_cmd('py.test v11tests/ -k "functional" -v --strict')
 	
 @task('run all v12 functional tests under current interpreter.')
 def v12_ft():